Move functions to the right place
* Move the register updating to a correct module
* move the req update to the apu
